requirements-dev.txt
# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0

# Linting & Formatting
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Type Stubs
boto3-stubs[essential]
types-boto3
sqlalchemy[mypy]
types-python-jose
types-aiohttp